+++ CHANGELOG.md
@@ -8,6 +8,7 @@ section below it for the last release. -->
 * Added a `Ada: Report Issue` command that opens the [VS Code Issue Reporter](https://code.visualstudio.com/docs/supporting/FAQ#_report-an-issue-with-a-vs-code-extension) with an extension-specific template
 * `Go to Definition` now jumps respectively on the `begin`, `private` and `body` keywords
   for subprograms, packages and tasks when clicking on the `is` keyword following their declarations
+* New refactoring: [Delete Entity](https://github.com/AdaCore/ada_language_server/blob/master/doc/refactoring_tools.md#delete-entity)
 
 ## 26.0.202507021